pyproject.toml
[project]
name = "cronos"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "twilio>=9.0.0",
    "anthropic>=0.40.0",
    "google-api-python-client>=2.140.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "apscheduler>=3.10.0",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
